refactor: API 폴더 구조를 클린 아키텍처 패턴으로 리팩토링
모든 route.ts 파일을 (adaptor) 폴더 구조로 이동하여 클린 아키텍처
원칙을 준수하도록 변경했습니다.

- app/api/duplicates/route.ts → app/api/duplicates/(adaptor)/route.ts
- app/api/signup/route.ts → app/api/signup/(adaptor)/route.ts
- .shared/docs/ → docs/ 폴더로 이동하여 문서 구조 정리
- 모든 테스트 파일의 import 경로 업데이트
- 문서 참조 경로 업데이트 (CLAUDE.md, GEMINI.md)
