+++ app/api/duplicates/(adaptor)/route.ts
