+++ app/api/signup/(adaptor)/route.ts
